Moved many things around: created lens.py to combine all types of lenses
edmund.py renamed edmundoptics.py thorlabs.py created olympus to put objectives.
